+++ src/main/java/com/truckcompany/example/TruckCompany/Services/TruckPartInventoryService.java
@@ -3,20 +3,29 @@ package com.truckcompany.example.TruckCompany.Services;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import com.truckcompany.example.TruckCompany.Domain.Category;
+import com.truckcompany.example.TruckCompany.Domain.Truck;
 
 import org.bson.types.ObjectId;
 import org.springframework.stereotype.Service;
 
 import com.truckcompany.example.TruckCompany.DataAbstraction.ITruckPartInventoryService;
 import com.truckcompany.example.TruckCompany.Domain.TruckPartInventory;
+import com.truckcompany.example.TruckCompany.Repositories.ICategoryRepository;
 import com.truckcompany.example.TruckCompany.Repositories.ITruckPartinventoryRepository;
+import com.truckcompany.example.TruckCompany.Repositories.ITruckRepository;
 
 @Service
 public class TruckPartInventoryService implements ITruckPartInventoryService {
     private final ITruckPartinventoryRepository truckPartInventoryRepository;
+    private final ICategoryRepository categoryRepository;
+    private final ITruckRepository truckRepository;
 
-    public TruckPartInventoryService(ITruckPartinventoryRepository truckPartInventoryRepository) {
+    public TruckPartInventoryService(ITruckPartinventoryRepository truckPartInventoryRepository,
+            ICategoryRepository categoryRepository, ITruckRepository truckRepository) {
         this.truckPartInventoryRepository = truckPartInventoryRepository;
+        this.categoryRepository = categoryRepository;
+        this.truckRepository = truckRepository;
     }
 
     @Override
@@ -26,18 +35,48 @@ public class TruckPartInventoryService implements ITruckPartInventoryService {
     }
 
     @Override
-    public void insert(TruckPartInventory item) {
-        truckPartInventoryRepository.save(item);
+    public Boolean insert(TruckPartInventory item) {
+        Category category = this.categoryRepository.findById(new ObjectId(item.getCategoryId())).orElse(null);
+        if(category == null)
+            return false;
+        
+        List<Truck> truck = this.truckRepository.findByIdIn(item.getTruckId());
+        if(truck.size() != item.getTruckId().size())
+            return false;
+        if(item.getName().length()<1)
+            return false;
+        return truckPartInventoryRepository.save(item).getId().length() > 0;
     }
 
     @Override
-    public void update(TruckPartInventory item) {
-        truckPartInventoryRepository.save(item);
+    public Boolean update(TruckPartInventory item) {
+        Category category = this.categoryRepository.findById(new ObjectId(item.getCategoryId())).orElse(null);
+        if(category == null)
+            return false;
+        
+        List<Truck> truck = this.truckRepository.findByIdIn(item.getTruckId());
+        if(truck.size() != item.getTruckId().size())
+            return false;
+        if(item.getName().length()<1)
+            return false;
+        TruckPartInventory itemToUpdate = this.truckPartInventoryRepository.findById(new ObjectId(item.getId())).orElse(null);
+        if(itemToUpdate == null)
+            return false;
+        itemToUpdate.setName(item.getName());
+        itemToUpdate.setTruckId(item.getTruckId());
+        itemToUpdate.setAmount(item.getAmount());
+        itemToUpdate.setCategoryId(item.getCategoryId());
+        this.truckPartInventoryRepository.save(itemToUpdate);
+        return true;
     }
 
     @Override
-    public void delete(String id) {
+    public Boolean delete(String id) {
+        TruckPartInventory itemToDelete = this.truckPartInventoryRepository.findById(new ObjectId(id)).orElse(null);
+        if(itemToDelete == null)
+            return false;
         truckPartInventoryRepository.deleteById(new ObjectId(id));
+        return true;
     }
 
     @Override
